Re-arrange Role inclusion to work around Moose limitations
The `requires()` can use attribute accessors to fullfill its needs.

Unfortunatly, the `has()` declaration must be done before the `with()`
calls.

More: if the attribute is defined inside a role, and required in another,
then this two roles must be composed with two distinct calls to `with()`:
the first call composes the provider role, and the second call composes
the requires role.

Signed-off-by: Pedro Melo <melo@simplicidade.org>